--- src/store/initial-state.ts
@@ -1,12 +1,0 @@
-import { AppState } from "./types";
-
-export const initialState: AppState = {
-  loading: false,
-  characters: {
-    list: []
-  },
-  filters: {
-    searchCriteria: "",
-    onlyFavorites: false
-  }
-};
