+++ src/types/index.ts
@@ -1,0 +1,14 @@
+type Comic = {
+  id: string;
+  title: string;
+  image: string;
+  year: string;
+};
+
+export type Character = {
+  id: string;
+  name: string;
+  image: string;
+  liked: boolean;
+  comics: Comic[];
+};
